Refacto leaks

+++ back/utils/ft_list_size.c
@@ -6,15 +6,15 @@
 /*   By: hhow-cho <hhow-cho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/23 00:12:02 by hhow-cho          #+#    #+#             */
-/*   Updated: 2019/06/23 00:12:10 by hhow-cho         ###   ########.fr       */
+/*   Updated: 2019/06/28 01:08:44 by hhow-cho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "shell.h"
 
-int ft_list_size(char **list)
+size_t ft_list_size(char **list)
 {
-    int size;
+    size_t size;
 
     size = 0;
     while (list[size])
